fix(metro-serializer-esbuild): polyfill `Symbol.asyncIterator` as necessary
