+++ .changeset/early-hounds-hide.md
@@ -1,0 +1,5 @@
+---
+"@rnx-kit/metro-serializer-esbuild": patch
+---
+
+esbuild [0.19.6](https://github.com/evanw/esbuild/releases/tag/v0.19.6) no longer throws when a `Symbol` is missing
