+++ packages/metro-serializer-esbuild/package.json
@@ -26,8 +26,7 @@
     "@rnx-kit/tools-react-native": "^1.3.4",
     "esbuild": "^0.19.0",
     "esbuild-plugin-lodash": "^1.2.0",
-    "fast-glob": "^3.2.7",
-    "semver": "^7.0.0"
+    "fast-glob": "^3.2.7"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0",
@@ -44,7 +43,6 @@
     "@rnx-kit/scripts": "*",
     "@types/jest": "^29.2.1",
     "@types/node": "^20.0.0",
-    "@types/semver": "^7.0.0",
     "eslint": "^8.0.0",
     "jest": "^29.2.1",
     "lodash-es": "^4.17.21",
